Verbose details for skipped model generation
It took me a lot of time to figure out why aren't the helpers being added to my classes, to realise that they were abstracts. I'm 100% its documented, but as my default debuging toolset was to add `-vvv` to the command, it didn't helped. 
This small fix would show why are we ignoring some files, instead of failing silently.